lab4/requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
python-dotenv==1.0.0
Werkzeug==2.3.4